# frontend/src/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Food Recommender API",
...
        "contact": {
            "name": "Food Recommender Team",
            "email": "support@foodrecommender.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        },
        {
            "url": "https://api.foodrecommender.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/restaurants/": {
            "get": {
                "tags": [
                    "restaurants"
                ],
                "summary": "Get Restaurant List",
                "description": "Get a list of restaurants from in-memory cache",
                "operationId": "get_restaurants_restaurants__get",
                "responses": {
                    "200": {
                        "description": "List of restaurants with basic information",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "restaurants"
                ],
                "summary": "Get Nearby Restaurants",
                "description": "Find restaurants near a specific location using DoorDash data",
                "operationId": "get_nearby_restaurants_restaurants__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NearbyRestaurantsRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Restaurants found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NearbyRestaurantsResponse"
                                },
                                "example": {
                                    "restaurants": [
                                        {
                                            "id": "restaurant_123",
                                            "name": "Tony's Pizza",
                                            "address": "123 Main St, City, State",
                                            "phone": "+1-555-0123",
                                            "rating": 4.5,
                                            "cuisine_tags": [
                                                "italian",
                                                "pizza"
                                            ],
                                            "image_url": "https://example.com/restaurant.jpg",
                                            "distance_miles": 0.8
                                        }
                                    ],
                                    "total_count": 25,
                                    "search_radius_miles": 5.0
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
        "/vapi/item-cart": {
            "post": {
                "tags": [
                    "vapi"
                ],
                "summary": "Get User Cart",
                "description": "Get Items currently in the user cart for the session.",
                "operationId": "get_user_cart_vapi_item_cart_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VapiRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/vapi/user_cart": {
            "post": {
                "tags": [
                    "vapi"
                ],
                "summary": "Add Item To Cart",
                "description": "Allow the user to add items to their cart.",
                "operationId": "add_item_to_cart_vapi_user_cart_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VapiRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/vapi/restaurant-by-id": {
            "post": {
                "tags": [
                    "vapi"
                ],
                "summary": "Get Restaurant Info",
                "description": "Get full information about restaurant - menu items, reviews, top items, other information",
                "operationId": "get_restaurant_info_vapi_restaurant_by_id_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VapiRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/vapi/restaurants": {
            "post": {
                "tags": [
                    "vapi"
                ],
                "summary": "Get Restaurant List",
                "description": "Fetch restaurants the user has access to order from",
                "operationId": "get_restaurant_list_vapi_restaurants_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VapiRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Message": {
                "properties": {
                    "toolCalls": {
                        "items": {
                            "$ref": "#/components/schemas/ToolCall"
                        },
                        "type": "array",
                        "title": "Toolcalls"
                    }
                },
                "type": "object",
                "required": [
                    "toolCalls"
                ],
                "title": "Message"
            },
            "NearbyRestaurantsRequest": {
                "properties": {
                    "latitude": {
                        "type": "number",
                        "title": "Latitude"
                    },
                    "longitude": {
                        "type": "number",
                        "title": "Longitude"
                    },
                    "radius_miles": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Radius Miles",
                        "default": 5.0
                    },
                    "limit": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Limit",
                        "default": 20
                    }
                },
                "type": "object",
                "required": [
                    "latitude",
                    "longitude"
                ],
                "title": "NearbyRestaurantsRequest"
            },
            "NearbyRestaurantsResponse": {
                "properties": {
                    "restaurants": {
                        "items": {
                            "$ref": "#/components/schemas/Restaurant"
                        },
                        "type": "array",
                        "title": "Restaurants"
                    },
                    "total_count": {
                        "type": "integer",
                        "title": "Total Count"
                    },
                    "search_location": {
                        "$ref": "#/components/schemas/RestaurantLocation"
                    }
                },
                "type": "object",
                "required": [
                    "restaurants",
                    "total_count",
                    "search_location"
                ],
                "title": "NearbyRestaurantsResponse"
            },
...
            "Restaurant": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "cuisine_tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Cuisine Tags",
                        "default": []
                    },
                    "location": {
                        "$ref": "#/components/schemas/RestaurantLocation"
                    },
                    "rating": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rating"
                    },
                    "price_range": {
                        "type": "string",
                        "title": "Price Range"
                    },
                    "image_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Url"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "location",
                    "price_range"
                ],
                "title": "Restaurant"
            },
            "RestaurantLocation": {
                "properties": {
                    "latitude": {
                        "type": "number",
                        "title": "Latitude"
                    },
                    "longitude": {
                        "type": "number",
                        "title": "Longitude"
                    },
                    "address": {
                        "type": "string",
                        "title": "Address"
                    },
                    "city": {
                        "type": "string",
                        "title": "City"
                    },
                    "state": {
                        "type": "string",
                        "title": "State"
                    },
                    "zip_code": {
                        "type": "string",
                        "title": "Zip Code"
                    }
                },
                "type": "object",
                "required": [
                    "latitude",
                    "longitude",
                    "address",
                    "city",
                    "state",
                    "zip_code"
                ],
                "title": "RestaurantLocation"
            },
...
            "ToolCall": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "function": {
                        "$ref": "#/components/schemas/ToolCallFunction"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "function"
                ],
                "title": "ToolCall"
            },
            "ToolCallFunction": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "arguments": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "object"
                            }
                        ],
                        "title": "Arguments"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "arguments"
                ],
                "title": "ToolCallFunction"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "VapiRequest": {
                "properties": {
                    "message": {
                        "$ref": "#/components/schemas/Message"
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "VapiRequest"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}